Adding distributor example
Sets up the following :

1. S3 bucket
2. Uploads packages to bucket
3. Create a `Package` type Document with Attachments pointing to s3
   bucket resources
4. Grants access to lambda that is creating the Document

Step 4 is important otherwise the Document never gets into Active status
and cannot be used due to the owner not having read permissions to the
s3 bucket.

Test/example also contains a section of commented out code that allows
for simply uncommenting it and commenting out the above section so that
you can validate update works as expected.